Cloud Functions/requirements.txt
flask
pandas
tensorflow==2.0.0b0
google-cloud-storage==1.16.1
numpy
sklearn
